All typescript errors in admin restaurant component have been resolved
