+++ pages/admin/restaurants/index.tsx
@@ -1,10 +1,11 @@
 import Sidebar from "../../../shared/components/admin/Sidebar/Sidebar";
 import AdminLayout from "../../../shared/components/admin/Layout/AdminLayout";
-import Adminresturant from "../../../shared/components/admin/AdminRestorant"
+import AdminRestaurant from "../../../shared/components/admin/AdminRestorant";
 import withAuth from "../../../shared/HOC/withAuth";
 import { serverSideTranslations } from 'next-i18next/serverSideTranslations';
 import { GetServerSideProps } from 'next';
 import { useTranslation } from "next-i18next";
+
 
  function Restaurants() {
     const { t } = useTranslation("common");
@@ -13,7 +14,7 @@ import { useTranslation } from "next-i18next";
         <>
             <AdminLayout>
                 
-                <Adminresturant />
+                <AdminRestaurant />
                 
             </AdminLayout>
         </>
